Multi-instance, limit number, and add tests
You can now supply multiple instances on the command line and it'll
actually work. It'll now output a table with instance type in it as
well.

The -n flag limits the number of line items to a specific count.

A testing framework is added. The main module is modified a bit to
support being run in test.